+++ Runtime/Scripts/INode.cs
@@ -1,9 +1,10 @@
 /*
+ * Copyright (c) 2026 Kokowolo. All Rights Reserved. 
  * Author(s): Kokowolo, Will Lacey
  * Date Created: December 19, 2024
  * 
  * Additional Comments:
- *      File Line Length: ~140
+ *		File Line Length: ~140
  */
 
 using System.Collections;
@@ -18,18 +19,6 @@ namespace Kokowolo.Pathfinding
         #region Properties
 
         public Node Node { get; }
-
-        #endregion
-        /*██████████████████████████████████████████████████████████*/
-    }
-    
-    public interface INode<T> : INode where T : INode
-    {
-        /*██████████████████████████████████████████████████████████*/
-        #region Properties
-
-        public new Node<T> Node { get; }
-        Node INode.Node => Node;
 
         #endregion
         /*██████████████████████████████████████████████████████████*/