+++ Runtime/Scripts/IPathfinding.cs
@@ -1,9 +1,10 @@
 /*
+ * Copyright (c) 2026 Kokowolo. All Rights Reserved. 
  * Author(s): Kokowolo, Will Lacey
  * Date Created: February 14, 2023
  * 
  * Additional Comments:
- *      File Line Length: 140
+ *		File Line Length: ~140
  */
 
 using System.Collections;
@@ -24,24 +25,24 @@ namespace Kokowolo.Pathfinding
         #region Functions
 
         // NOTE: this may change based on pathfinding configuration, i.e. a node's unit using its GetNeighbors() vs what's below
-        public List<Node> GetNeighborsFromNode(Node node)
+        public List<INode> GetNeighborsFromNode(INode iNode)
         {
-            return node.GetNeighbors();
+            return iNode.Node.GetNeighbors();
         }
 
-        public bool IsValidMoveBetweenNodes(Node start, Node end)
+        public bool IsValidMoveBetweenNodes(INode start, INode end)
         {
-            return start.HasNeighbor(end);
+            return start.Node.HasNeighbor(end);
         }
 
         // NOTE: this should get the distance/approximate between these two nodes
-        public int GetHeuristicCostBetweenNodes(Node start, Node end);
+        public int GetHeuristicCostBetweenNodes(INode start, INode end);
 
         // NOTE: this should get the actual distance between these two nodes
-        public int GetMoveCostBetweenNodes(Node start, Node end);
+        public int GetMoveCostBetweenNodes(INode start, INode end);
 
         // NOTE: this method is a hacky way of checking if the current path is too long; i.e. path.Distance > maxDistance;
-        public bool IsPathOutsideMovementRange(NodePath path)
+        public bool IsPathOutsideMovementRange(INodePath path)
         {
             return false;
         }