+++ Runtime/Scripts/NodeDebugObject.cs
@@ -1,4 +1,5 @@
 /*
+ * Copyright (c) 2026 Kokowolo. All Rights Reserved. 
  * Author(s): Kokowolo, Will Lacey
  * Date Created: August 22, 2022
  *